ctxutil: better go 1.21 compatibility
Release note: None
