+++ pkg/util/ctxutil/canceler_1_21_bazel.go
@@ -8,43 +8,33 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-//go:build !go1.20
+//go:build go1.21 && bazel
 
 package ctxutil
 
-import (
-	"context"
-	// Must import unsafe to enable linkname below.
-	_ "unsafe"
-)
+import "context"
 
 // A canceler is a context type that can be canceled directly. The
 // implementations are *cancelCtx and *timerCtx.
-// This interface definition is applicable to go1.19 or lower
+// This interface definition is applicable to go1.20 or higher
 type canceler interface {
-	cancel(removeFromParent bool, err error)
+	cancel(removeFromParent bool, err, cause error)
 	Done() <-chan struct{}
 }
 
-func (c *whenDone) cancel(removeFromParent bool, err error) {
-	if removeFromParent {
-		context_removeChild(c.Context, c)
-	}
+func (c *whenDone) cancel(removeFromParent bool, err, cause error) {
+	ctx := context.Context(&c.ExportedCancelCtx)
+	cx := ctx.(canceler)
+	cx.cancel(removeFromParent, err, cause)
 	c.notify()
 }
 
 type whenDone struct {
-	context.Context
+	context.ExportedCancelCtx
 	notify WhenDoneFunc
 }
 
 func makeWhenDone(parent context.Context, done WhenDoneFunc) {
-	c := &whenDone{Context: parent, notify: done}
-	context_propagateCancel(parent, c)
+	c := &whenDone{notify: done}
+	c.ExportedPropagateCancel(parent, c)
 }
-
-//go:linkname context_removeChild context.removeChild
-func context_removeChild(parent context.Context, child canceler)
-
-//go:linkname context_propagateCancel context.propagateCancel
-func context_propagateCancel(parent context.Context, child canceler)
