--- pkg/util/ctxutil/link_pre_1_21.go
@@ -1,24 +1,0 @@
-// Copyright 2023 The Cockroach Authors.
-//
-// Use of this software is governed by the Business Source License
-// included in the file licenses/BSL.txt.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0, included in the file
-// licenses/APL.txt.
-
-//go:build !go1.21
-
-package ctxutil
-
-import (
-	"context"
-	_ "unsafe" // Must import unsafe to enable gross hack below.
-)
-
-// Gross hack to access internal context function.  Sometimes, go makes things
-// SO much more difficult than it needs to.
-
-//go:linkname context_propagateCancel context.propagateCancel
-func context_propagateCancel(parent context.Context, child canceler)
